Move completion pipeline to separate file
