+++ GPTifier/include/run_completion.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+void create_chat_completion();
