+++ GPTifier/src/main.cpp
@@ -1,48 +1,10 @@
 #include "params.hpp"
-#include "prompts.hpp"
-#include "query.hpp"
-#include "requests.hpp"
-#include "responses.hpp"
-#include "typedefs.hpp"
-#include "utils.hpp"
+#include "run_completion.hpp"
 
 #include <iostream>
 #include <stdexcept>
 
 ::Params params;
-
-void create_chat_completion()
-{
-    ::str_request request = requests::build_chat_request();
-
-    if (request.empty())
-    {
-        throw std::runtime_error("Request is empty!");
-    }
-
-    requests::print_request(request);
-
-    QueryHandler q;
-    ::str_response response = q.run_query(request);
-
-    if (response.empty())
-    {
-        throw std::runtime_error("Response is empty!");
-    }
-
-    if (::params.dump.empty())
-    {
-        responses::print_chat_completion_response(response);
-        if (::params.enable_export)
-        {
-            responses::export_chat_completion_response(response);
-        }
-    }
-    else
-    {
-        responses::dump_chat_completion_response(response);
-    }
-}
 
 int main(int argc, char **argv)
 {
@@ -58,22 +20,6 @@ int main(int argc, char **argv)
     catch (std::runtime_error &e)
     {
         std::cerr << e.what() << std::endl;
-        return EXIT_FAILURE;
-    }
-
-    try
-    {
-        prompt::get_prompt();
-    }
-    catch (std::runtime_error &e)
-    {
-        std::cerr << e.what() << std::endl;
-        return EXIT_FAILURE;
-    }
-
-    if (::params.prompt.empty())
-    {
-        std::cerr << "Prompt cannot be empty" << std::endl;
         return EXIT_FAILURE;
     }
